Begin work on Chess.com / Lichess game fetching
I have started working on a feature where users can enter their Chess.com / Lichess.org username and pick out a game from their account. Website will in future extract the PGN of the chosen game and use it for them. Need to work on the backend soon, and then the app is pretty much finished!