Fixed more brilliant algorithm bugs
Fixed some bugs with the algorithm that determines whether or not a given piece is hanging. Sacrifices of a pawn - after a set of initial trades - was incorrectly considered a sacrifice of the original piece. Also added some ease-of-life stuff for developers including nodemon for automatic server refreshes and a report generation test.